pySideCar/requirements.txt
annotated-types==0.7.0
azure-core==1.34.0
azure-identity==1.23.0
azure-storage-blob==12.25.1
blis==0.7.11
catalogue==2.0.10
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
cloudpathlib==0.21.1
confection==0.1.5
conllu==6.0.0
cryptography==45.0.3
cymem==2.0.11
dotenv==0.9.9
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl#sha256=86cc141f63942d4b2c5fcee06630fd6f904788d2f0ab005cce45aadb8fb73889
idna==3.10
isodate==0.7.2
Jinja2==3.1.6
joblib==1.5.1
langcodes==3.5.0
language_data==1.3.0
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
msal==1.32.3
msal-extensions==1.3.1
murmurhash==1.0.13
nmslib-metabrainz==2.1.3
numpy==1.26.4
packaging==25.0
preshed==3.0.10
protobuf==6.31.1
psutil==7.0.0
pybind11==2.13.6
pycparser==2.22
pydantic==2.11.5
pydantic_core==2.33.2
Pygments==2.19.1
PyJWT==2.10.1
PyMuPDF==1.26.0
pymupdf4llm==0.0.24
pysbd==0.3.4
python-dotenv==1.1.0
regex==2024.11.6
requests==2.32.3
rich==14.0.0
scikit-learn==1.7.0
scipy==1.15.3
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
smart-open==7.1.0
spacy==3.7.5
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.1
temporalio==1.12.0
thinc==8.2.5
threadpoolctl==3.6.0
tiktoken==0.9.0
tqdm==4.67.1
typer==0.16.0
types-protobuf==6.30.2.20250516
typing-inspection==0.4.1
typing_extensions==4.13.2
urllib3==2.4.0
wasabi==1.1.3
weasel==0.4.1
wrapt==1.17.2
orjson==3.10.18
types-orjson==3.6.2
pytesseract==0.3.13
Pillow==10.4.0
